Restructing and adding the readme
