+++ oauth2-key-validation-adapter/Dockerfile
