+++ oauth2-key-validation-adapter/server.cer.pem
@@ -1,0 +1,25 @@
+Bag Attributes
+    friendlyName: wso2carbon
+    localKeyID: 54 69 6D 65 20 31 35 34 38 39 31 33 31 32 30 33 35 38 
+subject=/C=LK/ST=WS/L=SL/O=Home/OU=Home/CN=*.wso2.svc
+issuer=/C=LK/ST=WS/L=SL/O=Home/OU=Home/CN=*.wso2.svc
+-----BEGIN CERTIFICATE-----
+MIIDUzCCAjugAwIBAgIEN2LBVzANBgkqhkiG9w0BAQsFADBaMQswCQYDVQQGEwJM
+SzELMAkGA1UECBMCV1MxCzAJBgNVBAcTAlNMMQ0wCwYDVQQKEwRIb21lMQ0wCwYD
+VQQLEwRIb21lMRMwEQYDVQQDDAoqLndzbzIuc3ZjMB4XDTE5MDEzMDA5MTAxNloX
+DTE5MDQzMDA5MTAxNlowWjELMAkGA1UEBhMCTEsxCzAJBgNVBAgTAldTMQswCQYD
+VQQHEwJTTDENMAsGA1UEChMESG9tZTENMAsGA1UECxMESG9tZTETMBEGA1UEAwwK
+Ki53c28yLnN2YzCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAISUkno/
+o7FyItCpjqrCg6/CIYCv5zhDNmBgRsAVJ9YphZ8IrG4wOX+Fs4EDtK9EvJh6EfIe
+9TF6VY5l6NpQUSHeHAwqV1KFekzJSVrLF44GsIrdTXJd0fgEwu4w5dOW6uVzzYKV
+doNWk5XLq6ERgkFsOhLVBQn3JvWq/f53ypJoS46IA0cxmmFpW5cTweQ1eQLCRIhg
+hhzMLtlrQlUO7kF3MVkgIWkJQ6nVQk8R+OWHHUMvb7GOe1Lw3kOP/SeNn/0YUrs2
+q/l4OWcuIXf0hbkKRQ0Dijqj2uS2/tdDXmOfFo3+/lTrFGFnKGugRPeYMpzzKN4t
+KIHNLC9GqwSBEUMCAwEAAaMhMB8wHQYDVR0OBBYEFIcKVvxqy5m8eP1vWx2/KhuO
+vcKvMA0GCSqGSIb3DQEBCwUAA4IBAQAMQbKpkD9fBjnw5ctl/nCafgScksB+XznG
+LOhO97I+ZCFKpaaIXJixZpov3RNSbL+bf05QrP0WjApibfZndaKPLrlwl4H78M+x
+zScw1diutQgkqgcwSpAfdLxZ7+eSDchDF2R9Hg2NE+W2U6jU7oAdGqnh4r9AUzIb
+7m+Ye/NS60bfac22Fd4eHzV/rR/cLHBBnbBZd1QVbjOpPfETKB77DT/UZOmoUxG0
+v4qz75P8We5kccZwW8nb25G2tqT0Wgswpd/H9y3XTdqvMXwiCNfQ89BQx3obiA8n
+4qii45P+2vO01MCR9mw3dArD+jpATG/USQDQsbjrUXNw/FU+OK6e
+-----END CERTIFICATE-----
