+++ oauth2-key-validation-adapter/wso2/adapter-artifacts/attributes.yaml
