+++ oauth2-key-validation-adapter/wso2/adapter-artifacts/wso2_operator_cfg.yaml
