+++ oauth2-key-validation-adapter/wso2/cmd/main.go
