+++ oauth2-key-validation-adapter/wso2/config/adapter.wso2.config.pb.html
