+++ oauth2-key-validation-adapter/wso2/config/config.proto_descriptor
