+++ oauth2-key-validation-adapter/wso2/config/wso2.yaml
